requirements/base.txt
# QAToolBox 基础依赖
# 核心框架
Django==4.2.23
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==25.1
django-simple-captcha==0.6.2
crispy-bootstrap5==2025.6
django-ratelimit==4.1.0
django-ranged-response==0.2.0

# 数据处理
pandas==2.3.1
numpy==2.0.2
matplotlib==3.9.4
pyecharts==2.0.8

# 文件处理
Pillow==11.3.0
python-docx==1.2.0
pdfplumber==0.11.7
pdfminer.six==20250506
pypdfium2==4.30.0
xmind==1.2.0
reportlab==4.4.3

# PDF转换引擎 - 真实功能
PyMuPDF==1.25.0
pdf2docx==0.5.8
docx2pdf==0.1.8
python-pptx==0.6.23
openpyxl==3.1.5
xlrd==2.0.1
xlwt==1.3.0
pytesseract==0.3.10

# 网络请求
requests==2.32.4
urllib3==1.26.18
beautifulsoup4==4.13.4
lxml==6.0.0

# 浏览器自动化
selenium==4.20.0
webdriver-manager==4.0.1

# 异步任务
celery==5.3.4
redis==5.0.1

# 工具库
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
tenacity==9.1.2
prettytable==3.16.0
qrcode==7.4.2
yfinance==0.2.52

# 服务器
gunicorn==23.0.0

# WebSocket支持
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# 音频处理
pydub==0.25.1
mutagen==1.47.0
cryptography==42.0.5

# 数据库连接池
django-db-connection-pool==1.0.0
django-redis==5.4.0
redis==5.0.1

# 异步处理
celery==5.3.4
django-celery-beat==2.5.0

# 性能监控
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# 缓存优化
django-cacheops==7.2
